requirements.txt
google-adk
google-generativeai
google-genai
google-api-core
playwright
python-dotenv
pandas
opencv-python
numpy
graphviz
Pillow
typing_extensions
pydantic
